+++ R/info.R
@@ -82,6 +82,8 @@ mc_info_clean <- function(data) {
 #' * max_value - maximal recorded value
 #' * count_values - number of non NA records
 #' * count_na - number of NA records
+#' * height - height description of sensor
+#' * calibrated - logical value indicating whether the sensor is calibrated
 #' @export
 #' @examples
 #' mc_info(mc_data_example_agg)
@@ -116,7 +118,10 @@ mc_info <- function(data) {
                        min_value=purrr::map_dbl(item$sensors, function(x) function_with_check_empty(x$values, min)),
                        max_value=purrr::map_dbl(item$sensors, function(x) function_with_check_empty(x$values, max)),
                        count_values=purrr::map_int(item$sensors, function(x) length(x$values[!is.na(x$values)])),
-                       count_na=purrr::map_int(item$sensors, function(x) length(x$values[is.na(x$values)])))
+                       count_na=purrr::map_int(item$sensors, function(x) length(x$values[is.na(x$values)])),
+                       height=purrr::map_chr(item$sensors, ~ .x$metadata@height),
+                       calibrated=purrr::map_lgl(item$sensors, ~ .x$metadata@calibrated)
+                       )
     }
 
     prep_locality_function <- function(locality) {
@@ -347,5 +352,63 @@ mc_info_range <- function(data) {
     }
     result <- purrr::map_dfr(names(sensor_types), row_function)
 
+    as.data.frame(result)
+}
+
+#' Get calibration info table
+#'
+#' This function return data.frame with calibration parameter of sensors loaded by [myClim::mc_prep_calib_load()].
+#'
+#' @template param_myClim_object
+#' @return data.frame with columns:
+#' * locality_id - when provided by user then locality ID, when not provided identical with serial number
+#' * logger_name - name of logger in myClim object at the locality (e.g., "Thermo_1", "TMS_2")
+#' * sensor_name - sensor name either original (e.g., TMS_T1, T_C), or calculated/renamed (e.g., "TMS_T1_max", "my_sensor01")
+#' * datetime - date and time of calibration
+#' * cor_factor - correction factor applied to the sensor values
+#' * cor_slope - the slope of calibration curve
+#' @export
+#' @examples
+#' mc_info_calib(mc_data_example_clean)
+mc_info_calib <- function(data) {
+    is_raw_format <- .common_is_raw_format(data)
+
+    sensor_function <- function(locality_id, logger_name, sensor) {
+        count <- nrow(sensor$calibration)
+        if(count == 0) {
+            return(tibble::tibble())
+        }
+        result <- tibble::tibble(locality_id=rep(locality_id, count),
+                                 logger_name=rep(logger_name, count),
+                                 sensor_name=rep(sensor$metadata@name),
+                                 datetime=sensor$calibration$datetime,
+                                 cor_factor=sensor$calibration$cor_factor,
+                                 cor_slope=sensor$calibration$cor_slope)
+        return(result)
+    }
+
+    sensors_item_function <- function(locality_id, logger_name, item) {
+        count <- length(item$sensors)
+        purrr::pmap_dfr(list(locality_id=rep(locality_id, count),
+                             logger_name=rep(logger_name, count),
+                             sensor=item$sensors),
+                        sensor_function)
+    }
+
+    raw_locality_function <- function(locality) {
+        purrr::pmap_dfr(list(locality_id=locality$metadata@locality_id,
+                             logger_name=names(locality$loggers),
+                             item=locality$loggers),
+                        sensors_item_function)
+    }
+
+    if(is_raw_format) {
+        result <- purrr::map_dfr(data$localities, raw_locality_function)
+    } else {
+        result <- purrr::pmap_dfr(list(locality_id=names(data$localities),
+                                       logger_name=NA_character_,
+                                       item=data$localities),
+                                  sensors_item_function)
+    }
     as.data.frame(result)
 }